pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ada"
version = "0.1.0"
description = "Ada is a Matrix bot built with matrix.py framework. It provides various functionalities and can be extended with plugins."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matrix-python==1.4.5a0",
    "click==8.3.2"
]

[project.optional-dependencies]
dev = [
    "black",
    "pytest",
    "pytest-cov>=5.0.0",
    "flake8",
    "mypy",
    "types-PyYAML"
]

[tool.black]
line-length = 100
target-version = ["py312"]

[project.scripts]
ada = "bot.app:cli"

[tool.setuptools.packages.find]
include = ["bot*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
